after much work, things look mostly the same

+++ src/ui/control.h
@@ -46,7 +46,7 @@ struct AppState {
   const sf::RenderWindow &window;
   const Profiler &profiler;
 
-  float timeSince(const float event) const;
+  double timeSince(const double event) const;
 };
 
 /**
